okokokoihuhu

+++ lib/model/5_RWA_controller_RRR/rwa_payment_report_model.dart
@@ -4,9 +4,11 @@
 
 import 'dart:convert';
 
-RwaPaymentReportModel rwaPaymentReportModelFromJson(String str) => RwaPaymentReportModel.fromJson(json.decode(str));
+RwaPaymentReportModel rwaPaymentReportModelFromJson(String str) =>
+    RwaPaymentReportModel.fromJson(json.decode(str));
 
-String rwaPaymentReportModelToJson(RwaPaymentReportModel data) => json.encode(data.toJson());
+String rwaPaymentReportModelToJson(RwaPaymentReportModel data) =>
+    json.encode(data.toJson());
 
 class RwaPaymentReportModel {
   List<RwaPaymentReport>? rwaPaymentReport;
@@ -15,20 +17,26 @@ class RwaPaymentReportModel {
     this.rwaPaymentReport,
   });
 
-  factory RwaPaymentReportModel.fromJson(Map<String, dynamic> json) => RwaPaymentReportModel(
-    rwaPaymentReport: json["RWA_PaymentReport"] == null ? [] : List<RwaPaymentReport>.from(json["RWA_PaymentReport"]!.map((x) => RwaPaymentReport.fromJson(x))),
-  );
+  factory RwaPaymentReportModel.fromJson(Map<String, dynamic> json) =>
+      RwaPaymentReportModel(
+        rwaPaymentReport: json["RWA_PaymentReport"] == null
+            ? []
+            : List<RwaPaymentReport>.from(json["RWA_PaymentReport"]!
+                .map((x) => RwaPaymentReport.fromJson(x))),
+      );
 
   Map<String, dynamic> toJson() => {
-    "RWA_PaymentReport": rwaPaymentReport == null ? [] : List<dynamic>.from(rwaPaymentReport!.map((x) => x.toJson())),
-  };
+        "RWA_PaymentReport": rwaPaymentReport == null
+            ? []
+            : List<dynamic>.from(rwaPaymentReport!.map((x) => x.toJson())),
+      };
 }
 
 class RwaPaymentReport {
   int? id;
   String? patientName;
   String? bankName;
-  int? paidAmount;
+  double? paidAmount;
   String? paymentDate;
   String? paymentTime;
 
@@ -41,21 +49,22 @@ class RwaPaymentReport {
     this.paymentTime,
   });
 
-  factory RwaPaymentReport.fromJson(Map<String, dynamic> json) => RwaPaymentReport(
-    id: json["Id"],
-    patientName: json["PatientName"],
-    bankName: json["BankName"],
-    paidAmount: json["PaidAmount"],
-    paymentDate: json["PaymentDate"],
-    paymentTime: json["PaymentTime"],
-  );
+  factory RwaPaymentReport.fromJson(Map<String, dynamic> json) =>
+      RwaPaymentReport(
+        id: json["Id"],
+        patientName: json["PatientName"],
+        bankName: json["BankName"],
+        paidAmount: json["PaidAmount"],
+        paymentDate: json["PaymentDate"],
+        paymentTime: json["PaymentTime"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "Id": id,
-    "PatientName": patientName,
-    "BankName": bankName,
-    "PaidAmount": paidAmount,
-    "PaymentDate": paymentDate,
-    "PaymentTime": paymentTime,
-  };
+        "Id": id,
+        "PatientName": patientName,
+        "BankName": bankName,
+        "PaidAmount": paidAmount,
+        "PaymentDate": paymentDate,
+        "PaymentTime": paymentTime,
+      };
 }
